Introduce universal atomic non-volatile memory writes
Replaces direct `write` operations with a new `safe_save` method that employs a temporary file and rename strategy to ensure data integrity during writes to non-volatile memory.

This change also unifies the platform-specific `SRAM` and `data_file` implementations (previously in `l_mem.cpp/.h` and `w_mem.cpp/.h`) into the common `base_mem` module. This reduces code duplication, simplifies the build configuration, and streamlines memory management across different operating systems.
