+++ PAC/common/param_ex.h
@@ -60,15 +60,12 @@ class params_manager
             {
             C_TOTAL_PARAMS_SIZE =
 #ifdef PTUSA_TEST
-                1024 * 100,  ///< Общий размер памяти параметров для тестов.
+                1024 * 100,     ///< Общий размер памяти параметров для тестов.
 #else
-                1024 * 30,   ///< Общий размер памяти параметров.
+                1024 * 30,      ///< Общий размер памяти параметров.
 #endif
 
-            C_SYS_MEM_SIZE    = 10,          ///< Память для хранения CRC и т.д.
-
-            C_CRC_OFFSET      = 0,
-            C_LAST_IDX_OFFSET = 4
+            C_SYS_MEM_SIZE = 10 ///< Память для хранения CRC и т.д.
             };
 
         /// @brief Возвращает единственный экземпляр класса для работы с
